move k8s commands into one file
